+++ Data/Data/Invoice.cs
@@ -15,8 +15,8 @@ namespace MajoliFE.Data.Data
 		public DateTime CurrencyDate { get; set; }
 		public int CurrencyDateNumOfDays { get; set; }
 		public int CustomerId { get; set; }
-		public long BaseTotal { get; set; }
-		public long Total { get; set; }
+		public float BaseTotal { get; set; }
+		public float Total { get; set; }
 		public int PDV { get; set; }
 		public string Note { get; set; }
 		public bool IsPaid { get; set; }
